requirements.txt
Flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.0.3
python-dotenv==1.0.0
pytz==2023.3
lxml==4.9.3
schedule==1.2.0
